🛠️ Insert time range for top items (#146)
Inserting a time range (short, medium, and long term) for the top items (top songs, top artists, and top albums) to keep a history of the top listenings of the user. Yesterday’s short-term favourites are not the same as tomorrow’s top favourites; this enables to query the latest top items from the database.